feat: create image component
