+++ frontend/src/components/card/AssetCard.module.scss
